fix: added more precision to probing values

+++ client/src/Client.cpp
@@ -118,14 +118,14 @@ std::vector<std::string>* Client::fetchLog() {
 
 std::vector<std::string>* Client::probe(std::string address, int duration) {
     Prober prober = Prober();
-    std::vector<std::vector<int64_t>> *times = prober.stability(address, duration);
+    std::vector<std::vector<double>> *times = prober.stability(address, duration);
 
     int s = 1;
     std::vector<std::string> *probing = new std::vector<std::string>();
     // Format list to send to Redis
     for (auto const &perSecondValues : *times) {
         probing->push_back("SECOND$" + std::to_string(s));
-        for (int64_t const &value : perSecondValues) {
+        for (double const &value : perSecondValues) {
             probing->push_back(std::to_string(value));
         }
         s++;
